Implemented IAP to remove ads
- added store kit framework
- implemented logic to go through IAP process
- deliver functionality by flag in userdefaults to check whether or not
to show ads

+++ ObjectiveTrainerApp/ObjectiveTrainerApp/Menu/MenuViewController.m
@@ -45,12 +45,12 @@
 
 #pragma mark Table View Delegate Methods
 
-- (int)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
+- (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
 {
     return self.menuItems.count;
 }
 
-- (float)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
+- (CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
 {
     return 60.0;
 }
